+++ CHANGELOG.md
@@ -13,10 +13,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Create collection in database with all user, messages and chats
 - `/itt [-l LANG]` - Image to Text: Extract text from images
 - `/itt_lang` - Languages for ItT: Available languages for Image to Text
+- `/translate_image [TEXT] [-lf LANG] [-lt LANG]`  - Image to Text Translation: Extract text from images and translate it. `-lf` (default: detect, /itt_lang) language on image, to `-lt` (default: en, normal language codes) language.
 
 ### Changes
 - Fix command default options
 - Use Filters.all as default for MessageHandler
+- Yandex translate got new function for itself, it is used by the `/translate` and `/itt_translate` command.
 
 
 ## [1.1.2] - 2018-02-04